Add truncate support
There is no cross-platform truncate function in C++ or in Boost, so this adds a
callback which can be used to truncate a stream as required.  The tests use a
function to cut a stringstream, while the gamearch code uses the Linux
truncate() function to shorten the underlying archive file.  Some equivalents
will need to be added before it can be used on platforms other than Linux.