Add support for archives with the FAT at the end
Renamed insert/remove functions to preInsert and preRemove, and added
postInsert and postRemove functions so that actions could be performed both
before and now after a file's data has been added or removed.  This allows
archives with the FAT stored at the end to correctly update offsets and such.

The functions to set file sizes and offsets when editing FAT entries now also
have the delta size/offset supplied so that the function is aware of the amount
of change, which is important to keep track of where the trailing FAT has moved
to in these archive formats.