Add FixedArchive and Crystal Caves .EXE support
A FixedArchive makes fixed offsets within a file appear as files within an
archive.  This is slightly fake, but it does provide easy access to levels
and other data files stored in .exe and other files which are "fixed" and
can only have the data modified in-place.

Proof of concept for this is implemented in the form of the Crystal Caves
.exe file and the 17 levels contained within it.