Move decompression code out of Archive and into FilterType
Now Archives are no longer responsible for compressing and decompressing files
on the fly, making them simpler.  Instead they simply report which filter
(compression/encryption algorithm) is required, and it is the responsibility of
the caller to decompress or recompress the data as required.

This commit also includes a new example program "gamecomp" which provides a
command-line interface to the new filters, allowing standalone files to be
compressed and decompressed with any supported filter algorithm.  This paves
the way for additional algorithms to be implemented which are not necessarily
associated with archive files (e.g. Zone 66 files which are not grouped
together in an archive but are still compressed.)