New archfile stream for providing direct stream::sub access to FATArchive files
This new class inherits from stream::sub, but provides the offset and length
directly from the FATEntry fields.  This allows all open streams in an archive
to be moved around automatically as other files are added and removed.

It also holds only one strong reference to the FATArchive instance, so that it
can call resize() as needed.  This also means the archive can be closed but the
file can still be used successfully.