tests: Add ability to skip isInstance tests on new archives that would fail
Some archives have no signature, so newly created ones can be difficult to
detect, e.g. if they are zero bytes long, or the archive produced in the test
doesn't match any known file sizes (where the only available method of
detection is to check against a list of known file sizes.)  So this allows
those tests to be skipped.