New filter for handing compressed files with UINT32LE decompressed-size header
This filter just crops the data at the indicated size when reading, and inserts
the size when writing.  It is useful to avoid handling this case in each
compression filter, as well as ensuring files don't have trailing junk when
there is no "end of data" marker in the compressed data.