ddave-rle: Use decomp_size filter to move the data length into the filter
Conceptually the decompressed size is part of the file format itself, so this
better fits with that idea.  It also simplifies the format handler a little, as
it no longer has to shorten the file by four bytes and update the file size
through a callback.

The callback is still there for the moment however, until a better way can
be implemented for the decompressed size to be reported.