Remove fmt-mni-czone
Duke Nukem II CZone tilesets have been moved back to libgamegraphics.

Originally it seemed better treating the CZone as an archive containing three
files (tile attributes and two tilesets) however not only did this end up being
a poor-man's archive (fixed number of files, can't resize, etc.) it turned out
to be problematic for loading maps.  Each map stores the name of the CZone
tileset which is returned by libgamemaps, but there's no way to indicate that
the CZone file should be opened as an archive and only the last two files be
read as tilesets for rendering the map.  Instead, it's much easier to just
report the CZone file as the tileset, and have libgamegraphics load it,
treating it as one tileset containing two sub-tilesets.

The original motivation behind moving CZone support to libgamearchive in the
first place was handling of the non-tileset "tileinfo" data, however it seems
that this may work best by being presented as part of the tileset after all.